+++ Cargo.toml
@@ -84,6 +84,7 @@ members = [
     "crates/storage/db/",
     "crates/storage/db-api/",
     "crates/storage/db-common",
+    "crates/storage/db-mdbx",
     "crates/storage/errors/",
     "crates/storage/libmdbx-rs/",
     "crates/storage/libmdbx-rs/mdbx-sys/",
@@ -116,7 +117,7 @@ members = [
     "examples/rpc-db/",
     "examples/txpool-tracing/",
     "testing/ef-tests/",
-    "testing/testing-utils"
+    "testing/testing-utils",
 ]
 default-members = ["bin/reth"]
 
@@ -248,6 +249,7 @@ reth-consensus-common = { path = "crates/consensus/common" }
 reth-db = { path = "crates/storage/db", default-features = false }
 reth-db-api = { path = "crates/storage/db-api" }
 reth-db-common = { path = "crates/storage/db-common" }
+reth-db-mdbx = { path = "crates/storage/db-mdbx" }
 reth-discv4 = { path = "crates/net/discv4" }
 reth-discv5 = { path = "crates/net/discv5" }
 reth-dns-discovery = { path = "crates/net/dns" }
@@ -325,8 +327,14 @@ reth-trie = { path = "crates/trie/trie" }
 reth-trie-parallel = { path = "crates/trie/parallel" }
 
 # revm
-revm = { version = "9.0.0", features = [ "std", "secp256k1", "blst", ], default-features = false }
-revm-primitives = { version = "4.0.0", features = [ "std", ], default-features = false }
+revm = { version = "9.0.0", features = [
+    "std",
+    "secp256k1",
+    "blst",
+], default-features = false }
+revm-primitives = { version = "4.0.0", features = [
+    "std",
+], default-features = false }
 revm-inspectors = { git = "https://github.com/paradigmxyz/evm-inspectors", rev = "53aa2b2" }
 
 # eth
