+++ crates/storage/db-mdbx/src/mdbx/mod.rs
@@ -1,11 +1,24 @@
 //! Bindings for [MDBX](https://libmdbx.dqdkfa.ru/).
 
 pub use crate::implementation::mdbx::*;
-pub use reth_libmdbx::*;
-
 use crate::is_database_empty;
 use eyre::Context;
+pub use reth_libmdbx::*;
 use std::path::Path;
+
+pub mod cursor;
+pub mod tx;
+
+const GIGABYTE: usize = 1024 * 1024 * 1024;
+const TERABYTE: usize = GIGABYTE * 1024;
+
+/// MDBX allows up to 32767 readers (`MDBX_READERS_LIMIT`), but we limit it to slightly below that
+const DEFAULT_MAX_READERS: u64 = 32_000;
+
+/// Space that a read-only transaction can occupy until the warning is emitted.
+/// See [`reth_libmdbx::EnvironmentBuilder::set_handle_slow_readers`] for more information.
+#[cfg(not(windows))]
+const MAX_SAFE_READER_SPACE: usize = 10 * GIGABYTE;
 
 /// Creates a new database at the specified path if it doesn't exist. Does NOT create tables. Check
 /// [`init_db`].
