--- crates/storage/db/src/implementation/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "mdbx")]
-pub(crate) mod mdbx;
